fix: some react confirm modals
